requirements.txt
requests==2.31.0
python-dotenv==1.0.1
ratelimit==2.2.1 
tqdm==4.67.1